♻️ M9K6C0 cli: decompose preflight report
