+++ .agentplane/tasks/202605290043-M9K6C0/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290043-M9K6C0`
+Title: Preflight report command decomposition
+Canonical task record: `.agentplane/tasks/202605290043-M9K6C0/README.md`
+
+## Summary
+
+Preflight report command decomposition
+
+Decompose packages/agentplane/src/cli/run-cli/commands/core/preflight-report.ts into focused preflight report modules while preserving CLI behavior and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/cli/run-cli/commands/core/preflight-report.ts into focused preflight report modules while preserving CLI behavior and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Preflight report command decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:43:29.282Z
+- Branch: task/202605290043-M9K6C0/preflight-report-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
