+++ .agentplane/tasks/202605290043-M9K6C0/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 M9K6C0 task: Preflight report command decomposition [202605290043-M9K6C0]
