+++ .agentplane/tasks/202605290043-M9K6C0/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T00:43:29.282Z
+
+## Task
+
+- Task: `202605290043-M9K6C0`
+- Title: Preflight report command decomposition
+- Status: DOING
+- Branch: `task/202605290043-M9K6C0/preflight-report-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290043-M9K6C0/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:43:29.282Z
+- Branch: task/202605290043-M9K6C0/preflight-report-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
